gRPC/pyproject.toml
[project]
name = "grpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grpcio-tools>=1.75.0",
]
